# specs/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Cafe
  license:
    name: MIT
servers:
  - url: http://cafe.swagger.io/v1
paths:
  /cafes:
    get:
      summary: List all cafes
      operationId: listCafes
      tags:
        - cafes
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            maximum: 100
            format: int32
      responses:
        '200':
          description: A paged array of cafes
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cafes"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a cafe
      operationId: createCafe
      tags:
        - cafe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cafe"
      responses:
        '201':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cafes/{cafeId}:
    get:
      summary: Info for a specific cafe
      operationId: showCafeById
      tags:
        - cafes
      parameters:
        - name: cafeId
          in: path
          required: true
          description: The id of the cafe to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cafe"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Cafe:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        tag:
          type: string
    Cafes:
      type: array
      maxItems: 100
      items:
        $ref: "#/components/schemas/Cafe"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string